.history/requirements_20260206154327.txt
Flask
Flask-SQLAlchemy
Flask-Login
qrcode[pil]
xhtml2pdf
requests 